PCU: Fix issue when process terminated before startup completes
*) A state change was not send out when a process terminates before the
#start method returned. In that case only STOPPED was send out but the
START was missing.

*) Adding a completion hook to a CompletableFuture triggers immediately
when the future is already done. Thus we do not need to manually check
that situation. This solves the issue that two state changes where
performed within the startup (START, STOPPED) task. Now the exit hook is
executed after startup is completed.

*) A grace period of 250ms is now set for all PCU tests so that a
process is given some time to terminate. Should fix issues in the
InstanceProcessControllerTests where no actual waiting is performed.

Request: MO
Change-Id: Ia427bbaf5b23e523922f080e39ea4fb7a3a2a304