Fix PKIX path building problems
HttpsUrlConnection's getDefaultSSLSocketFactory() is unfortunately not
thread safe. This alone is mostly not a problem, but Jersey's
HttpUrlConnector checks for instance equality before setting our
SSLContext (to avoid the costs of the SecurityManager) - which fails if
a second thread re-calculates a 'new' default SSLSocketFactory.

The workaround is to call (and thus initialize)
getDefaultSSLSocketFactory() in a static (single threaded) context once
on startup to avoid race conditions later

Bug: DCS-417
Change-Id: Ic5fd6a4226975c09538ca3c6467dd1a5a5f9ab3c