Windows Installer: Fix issues with path separators in ZIP
*) The ZIP file contained mixed (Windows, Linux) path separator
characters and thus we failed to extract all files. This is now fixed in
Java and in C#. We normalize the path on both sides now so that this
should not happen any more.

*) The installer now creates a start menu entry so that the user finds
the application at a common place. We create a top-level entry in the
'Programs' folder of the current user.

Change-Id: Ib0ec4c911af7e8a5e99b2f10e1ef44e114cea563