Check server certificate before establishing connection in installer
Linux installer: download and check the server certificate before
downloading files. This is done manually, as cURL cannot check the
certificate without ALSO checking the hostname (which is not allowed for
our use case as the hostname can be different depending on the network
setup and location of the client).

This mechanism protects the "normal" user against server spoofing
sufficiently. The mechanism can be avoided by the user however by
editing the installer and removing the check - but this would be the
same with cURL as well :)

Bug: DCS-520
Change-Id: I824e1ad4e5a274b10ea59ccb28d3857b36b6647c