Further locking to prevent PruneOperation from missing marker databases
There was still a tiny whole where an operation could have deleted a
marker database while the PruneOperation was reading the list of objects
from the actual ObjectDatabase. In this case it would miss the manifest
(not there yet), read all objects, and fail to protect them (marker
database deleted).

Request: DCS-691
Change-Id: I5e9865387925e3a7b13d1be3a829e693630dbf69