Remote Master: no longer store with instance configuration
 * STANDALONE, LOCAL will always use "self" as target. If an instance
   is in the hive, it is obviously meant for it.
 * CENTRAL will determine the target master later on, unsing meta
   manifests attached to the instance (root).
 * TOOL will always receive the remote to connect to as parameters.
 * SLAVE is not allowed to connect to any master.

The master configuration has been removed from the add/edit instance
dialog completely, as it is not longer required in this form. Later on,
we will need to be able to choose from available targets on the central
server (only!). This information will then be persisted in meta manifest
next to the actual instance configuration pushed to the target (the
target does not need this information).

Request: DCS-719
Request: DCS-720
Change-Id: Ib451ae5a842db902113fccc0b06efa1db77d47cf