Client Launcher: New installation / update behavior
*) The installation as well as the update is now done without involving
the DCU. Using the DCU had the drawback that an update to any of the
configured processes leads to an update of the client application (even
if just a server process was changed).

*) The entire application as well as its dependencies are fetched if
they are not yet existing on the client. The files are stored in a
pooled directory to avoid duplication.

*) The actual command line to launch the process is always fetched from
the server. This information is NOT persisted. This has the advantage
that configuration changes do not require an update of the installed
application. Only if the product version is changed an update is
required.

*) The software that is required by an application is tracked using a
new manifest. The cleanup removes each software that is not in use any
more by any of the installed client applications.

Request: DCS-684
Change-Id: I515b0062aec745624d027d9b7aa9469c5c758b41