Support multiple clients launchers
*) The launcher now checks if the application to launch requires an
older launcher version. If so then the (older) launcher will be
downloaded and installed.

*) Older launchers are installed in a version-controlled directory
inside of the BDeploy root directory (BDEPLOY_HOME\bdeploy-v1.3.1)
Each version contains the launcher, the hive as well as all apps.
Nothing is shared between launcher versions. This avoids unintended
side-effects.

*) The actual launching is delegated to the older launcher. This ensures
that changes that we make in newer versions do not break older
launchers.

*) Launchers that are not used any more - because the server has been
updated - will automatically be removed. Thus the system will cleanup
everything that is not required any more.

*) Removing an application is also delegated to the corresponding
launcher that is in control of the application. If all apps of an older
launcher are removed the entire launcher is removed.

Bug: DCS-888
Request: DCS-875
Change-Id: Ia030befdd8f50ec05dc2d331e54406d0183eb6c5