Implement system user migration
Previously "system" tokens were issued using the 'user.name' system
property instead of a constant. Since user names are now validated in
the master, these old tokens are no longer accepted (1.3.1 or older was
used to init the BDeploy root).

This migration detects such an old token and regenerates it. This token
is used when attaching a managed server and whenever communicating with
itself (i.e. every master is also a slave).

Request: MO
Change-Id: I3310bb80086b8f6699196da3bbb7c78ed2f6a240