PCU: Startup/ Shutdown order of processes
*) Processes are now started / stopped in a well defined order. The
order is taken from the instance configuration. Thus by dragging
applications around the user can influence the startup/shutdown order.
For stopping the order is simply reversed so that the first app is
stopped as last one.

*) Limitation: The order is just guaranteed within one node. There is no
way to define the order when processes are distributed across nodes.
This might be added at a later point if required.

Bug: DCS-937
Change-Id: If43385d0c1e2a51c3c54ef65ee32bc6f0fcb4b6d