PCU: Better feedback for the user when stopping an instance
*) Stopping an entire instance lacked an appropriate feedback for the
user that initiated the shutdown. All processes remained running and
nothing happens. The processes just disappeared one after each other.

*) To solve that a new process state 'RUNNING_STOP_PLANNED' has been
introduced. This indicates that the processes will be stopped in the
near future. Thus the user gets an immediate feedback that something is
in happening after he clicked the stop button.

Bug: DCS-937
Change-Id: Id1b9d37429a35c6e21da1be89a163db0e44999e4