Provide hints about autofill to the browser
According to the official spec at [1] this should prevent the browser
from 1. storing and 2. autofilling unwanted information into application
configuration input controls.

The combination of a propper name (which prevents "guessing" by the
browser) and a set autocomplete attribute on the controls form-owner
should do the trick for any browser supporting the spec.

[1] https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#autofill

Bug: DCS-1001
Change-Id: I4c5d9cf029cbf8f0e086457a1ba861d88be4b38b