Fix issues with multiple launchers running at the same time
 * Allow only one launcher process to do:
  * Updates of the launcher
  * Installation of an application
  * Removal of an application

This is enforced using a global lock on the directory. Applying a
launcher update will use an additional marker to signal other launchers
that an update is in progress. This marker is valid for 60 seconds only.
The marker will be removed once the update is done by the current shell
scripts, but not (yet) on windows - needs to be done in the C#
application.

Bug: DCS-1071
Bug: DCS-1029
Change-Id: I88775e9cd3a53d5d716e93eab7dcd19c6d5f5c99