Stop command handling: await grace period after stop command
Changes the way the stop command is treated. With this change, the stop
command can either be synchronous or asynchronous. Before it was only
possible to use synchronous commands, as the main process was destroyed
immediately after the stop command exited. Now the grace period is
awaited in case an asynchronous trigger causes the main process to exit
within that period.

Request: MO
Change-Id: Ia72eb54f8c895b19e1a117a5e6a953206db30ce5