Launcher: Fix issues that cleaning can result in corruption
*) Two applications that are simultaneous updated could corrupt the pool
when one tries to delete some files that are still locked by the other
application. This results in a pooled application where some but not all
files are deleted.

*) To avoid that we now try to rename the folder before deleting the
content. Moving fails if there are still some files locked. This
guarantees that we can remove either all files or no files at all.

Request: MO
Change-Id: I920da3d7c2cf753e09cd990fe641c945b24b46a8