Disable coverage aggregation on Windows
Windows builds are simply too slow on GitHub to have a chance for
coverage aggregation to ever finish. Frequent timeouts are the result

The application itself is still run with instrumented code which *would*
track coverage. The data is simply not evaluated and written to disc by
the cypress plugin.

Request: MO
Change-Id: I5c176e29a4117d755545ec83d3548e57eef5f783