PCU: Fixed an issue that a process could be stuck in a certain state
*) Scheduling a start command while the PCU is executing the stop
command stopped could lead to the situation that the application remains
forever in 'RUNNING_STOP_PLANNED'.

*) The start command waits until the stop completed. But the stop
command also triggers an exit event which waits until the start command
is done. This then caused an exception when trying to get the exit code
of a running application. This messed up the internal state and the app
remains forever in the same state :(

*) The fix is that the exit runnable checks if the process handle
changed in the meantime. If so then it simply does nothing.

Bug: DCS-1214
Change-Id: If58d9044d0bca40ad578a4de16a393309d81d335