Log file browser and viewer
This is basically a stripped down copy of the data files browser for
now. It can list/display/download log files from all nodes. The plan is
to extend the according UI component with the functionality to edit the
logging config online. The required backend is already part of this
commit.

Request: DCS-1228
Change-Id: I33d4bfb9d38fd9614d5bc3247793e4000dbdb6c8