Launcher: Improvements and fixes
*) The launcher now supports the --updateOnly flag that ensures that the
launcher as well as the application is up-2-date. After applying any
updates the launcher terminates again.

*) When the C# installer is running in the unattended mode then the
launcher is started with the --updateOnly flag so that also the
application is installed. This makes automated deployments easier as
after the installer is finished the launcher as well as the target
application is installed on the system.

*) The launcher now passes the --customizeArgs flag to the older
launcher so that customization is also working with different version.

Bug: DCS-1366
Bug: DCS-1335
Change-Id: I5a5be2ff9d7982dfcd87a0b86267339730231098