Use SYNC OpenOption when writing critical file content
This forces writes to both file content and file metadata to be finished
before the operation(s) return. When *not* using SYNC, writing may
return before content and metadata are actually written to disc. This in
turn can result in data loss on power outage or disc space issues.

The according Java documentation states that "critical" data should be
written in this mode. In our case nearly every single (non-unit-test)
file is critical, so we generously use this mode now to write almost
everything which is not purely temporary.

Request: DCS-1368
Change-Id: If38d73e022ee1d2796622b1a8927759ae56e4476