Launcher: Support passing additional command line arguments
*) When launching an application using a .bdeploy file the user can now
specify additional command line arguments that are then passed to the
application.

*) We need to distinguish between arguments for the internal launcher
and arguments designated for the application to be launched. Therefore
we decided that ALL arguments after the -- switch are passed to the
launched application. All other arguments before are treated as
arguments for the launcher. They will not be forwarded.

*) Additional arguments having a space must be quoted. Otherwise they
will be passed as two individual arguments and not as single one.

Bug: DCS-1295
Change-Id: I3442c69cd395ed3e072b64128faf0048c56332a1