Fix grouped sending of activities to clients
All activities matching the requested scope of each client are now
grouped into a single update message.

The only exception is when a scope is no longer present (removed due to
all activities in this scope done). In this case an update with an empty
message list is send to make sure *some* update is sent to the client in
case this was the only subscribed scope. In this case another update
will immediately afterwards set the new activities. This will result in
a slight flickering in the view in some cases.

Request: DCS-1352
Change-Id: I667f99407446df8a7a54518dc8d2facf2798c29b