Fix dirty handling in panels opened on non-dirtyable primaries
If the primary dialog is not dirtyable, the DirtyDialogGuard was never
called, and thus no confirmation on the panel was performed in case the
primary outlet performed a navigation.

Primary dialogs now must *always* have the DirtyDialogGuard, even if
they themselves are not dirtyable, to be able to catch dirty panels.

Request: MO
Change-Id: If36aeb5aba30703fbb4d3ef2f8e22ed6db8e0e0f