Firefox interprets certain CSS *very* different from chrome (& co.)
The client usage graph was way too wide, as the flex seems to be applied
in a different stage of the layouting. Thus we set the flex basis to
very low, so it later expands, instead of being required to shrink
(which does not happen correctly).

The width of columns is *super* tricky. There are various algorithms
that try to distribute "remaining" widths intelligently, which results
in a completely f-ed up picture with FireFox. width: 0 in chrome is the
same as width: auto in FireFox. width: 0 in FF will result in FF
distributing *remaining* width among *all* columns (as all have a width
set). This will make fixed-width columns wider than expected, and
non-width-set columns way narrower than expected. Setting *some* width
on all columns is also super important, as otherwise layout breaks
completely. Setting width to 'auto' fortunately seems to work in all
browsers as expected.

Request: MO
Change-Id: I1f364c6843bdabb565c871c5d045f2caf8f12861