Fix wrong order of objects in pushes and fetches
This causes a complete mess in case a push or fetch fails for *any*
reason. Since leafs are *not* correctly sent before trees, this will
cause following retries and manual push/fetch retries to fail due to
trees being present where leafs are missing (which is undetected for
performance reasons).

Bug: DCS-1388
Change-Id: Ice4cf85f3c22273b6a71632a18aa5e71fbb992f3