Launcher: Harden against broken software manifests
*) The application browser now does not crash anymore as soon as their
is one or more broken software manifests in the hive. Such broken
manifests are simply ignored. They will not show up in the given
list.

*) Errors during the cleanup phase that the launcher is doing after the
application startup are now simply logged and no exception is raised
anymore. Showing an error message after we successfully started the
application is confusing for the user.

*) When an exception is raised during the application clean we
automatically execute a 'fix errors' and 'prune' command to get the hive
back in an normal state.

*) 'Fix Errors' and 'Prune' are available as actions in the 'Browser'.
This actions are operating on all hives that are available locally.

Bug: DCS-1394
Change-Id: Ifb1457d4d3faca3fbad868a4a23a7a4786b07d6b