Fix recursive cross-scope variable resolving.
E.g. a config file which references {{V:app:param.1}}. param.1 in
application app has value {{V:param.2}} pointing to another parameter in
the same application. In this case the resolver scope is the config
file, not the application, which would not work. The {{V:param.2}} needs
to be re-scoped to the application scope in this case.

Bug: DCS-1032
Change-Id: I90cc6e015f89bec7af58280b863572f67ec73754