Introduce warning when WebSockets have a problem.
BDeploy 4.0 *heavily* relies on WebSockets for proper operation. If they
have a problem (i.e. proxy, vpn, firewalls, browser, you name it) data
will not be received from the server, and the user needs to manually
refresh to get current data.

This is not a proper fix for any issue that arises indirectly from
broken websockets, but it will at least give *us* a hint what is going
on if we see pages with outdated data.

Bug: WPL-137
Change-Id: Ieb2d3995097d1c6ff77f20639f657cdf4baf6308