Avoid exceptions when updating search during navigation
The issue arises if components are destroyed during navigation, and the
search is updated. This happens in an asynchonous way and *may* or may
not trigger an exception. This in turn brings the complete webapp to a
halt.

We may be able to find a better fix later, for now we simply log and
ignore the issue (as the component is being destroyed anyway).

Bug: WPL-441
Change-Id: I5c546a434edd66209d7398bf2393addf585248ba