Fix data files add dialog to not "stop saving" when asking for replace
This made UI tests flaky, as they wait for the spinner to disappear from
the DOM, and the spinner in this case was disabled while asking the user
whether to save. The test continued, the user confirmed, but on slower
machines this takes so long (still only milliseconds) that cypress will
already have queried the DOM for the element to check - which is still
the wrong one from the old DOM before an update happens to the table.

Request: MO
Change-Id: Ieac4f3ecfa6d601c6051ff76602c9fcb8922e3c7