Make sure to not show wrong/outdated data to users
The data is still in the services when changing to *another* group/repo,
however it actually belongs to the *previous* group/repo. For ~100ms,
the user is able to see the previous data when entering another
group/repo. If fast enough with clicking, he even gets into invalid
situations in the UI.

Bug: WPL-585
Change-Id: Ie4d7f7080263d0f300b9887c41331a2262b27bf2