Record instance version deletion in the instance manifest history
The deletion event is visible is the same instance version number is
created again, but not before. This is only to have a chance that *if* a
version is re-created (and shown twice), why this is the case. Also this
event records the user and the time when the deletion happened.

Bug: WPL-778
Change-Id: I5efdab61c618bfed4c6bb878771e36532cef7374