Fix for corner case which could cause SSL contexts to break
In case applications are started and probes are contacted *before*
JerseyClientFactory is initialized, there *may* be a corner case which
causes the SSL context to be initialized wrong. We counteracted this
corner case before in JerseyClientFactory, but CommonEndpointHelper
needs the same workaround to assure that it does not matter which of the
two is initialized first.

Request: MO
Change-Id: I1767dcb34ca6ab96a59f6c6d2d3897d029be1343