Make sure to prune after a failed transfer to not leave damaged objs
In case manifest consistency is not ok, damaged objects may have lead to
that. Deleting the manifests is not enough as a damaged tree may stay
around and not be removed. These objects are unreferenced after manifest
deletion, so a prune will remove them.

This does not *prevent* errors per se, but it will put the hive in a
state where the operation has a chance to succeed on next retry.

Bug: WPL-1519
Change-Id: I4fc630090e8d11a90f6fa2551e38a6ba9c31eb86