Properly normalize paths for segmented manifest list cache
This issue led to outdated manifest caches, up to 10 minutes. This in
turn lead to the prune at the end of the nightly cleanup to *not* see
manifests created in the last 10 minutes (e.g. *during* cleanup) and
thus pruning objects which are required by those!

Also added a unit test verifying this bad behavior and fixed by adding
the additional .normalize() on paths to force them to be equal.

Bug: WPL-2473
Change-Id: I7240fac37afce28745e225aa2d87e0e670509349