Changed idle handling to no longer attempt to (re)open the connection
The new philosophy of the idle thread is to simply keep retrying to idle
until it is interrupted from the outside or throws an unexpected
exception. It DOES NOT try to establish a connection or open a folder by
itself. If the idleing fails due to one of those circumstances, the
method simply retries after a given timeout. Other threads are
responsible for keeping the connection alive.

Also removed a redundant return at the end of the idle method

Request:WPL-2135
Change-Id: Ia48f947b22439287646ed8de30a2734355b2039c