Made the punctuation and wording of base cli tools consistent
In most cases this just means that the dot at the end was removed. Also
removed a duplicate whitespace and fixed a typo where a word was
incorrectly started with an uppercase letter.

This commit only handles the cli tools which are present at the default
overview, e.g. NOT tools like the BrowserTool.

Request: None
Change-Id: I3a63ca1334c597c6130f4c17b59f5872fa3ab448