Use system look and feel only on Windows, prevent Linux dark theme
On linux setting legacy themes with high contrast or dark theme will
render only parts of the application correctly, thus we need to avoid
this by setting the cross platform look and feel, which does not react
to system themes at all.

The cross platform look and feel looks quite oldschool and not as nice
as the native one, but dark theme will not destroy it.

Bug: WPL-2899
Change-Id: If3d29ce85acb845ed53cffe4a6431797dc3edc10