Fixed -op/-vv handling + general ToolBase Improvements
Before this change the behavior of the -op arg depended on whether it
was called before or after the -vv arg.

Now all (relevant) args are only processed once instead of overwriting
each others calculations repeatedly. Also, argument order in general
matters 'less'. The behavior is as follows:
* For all verbosity args (-q|-v|-vv), the last one wins
* For the data format args, the last one wins (unchanged)
* If multiple -o args are given only the last one is processed and used
* If multiple -op args are given only the last one is processed and used

The order of the args in the switch is now the same as the order of the
args in the summary.

In general the logic has been a little simplified. E.g. no more
nullchecks are necessary to determine if an argument has been provided.

Finally, the JavaDoc of the ToolDefaultVerbose annotation got updated to be a little more specific.

Request:None
Change-Id: I75a83cf714ade48692f38bdccb67fa8057fa14bf