Removed duplicate check of same value
The spread operator throws an exception if it is used on an undefined
object. However, writing it like ...(obj && obj) does not safeguard
against this exception.
The exception is thrown because the spread operator only works on
iterables, and undefined is not iterable. However, (undefined &&
undefined) just resolves to undefined which then throws the same error.

For this very reason I also removed the ? operator in the mappings. If
either g or r would ever be nullish, the operator would resolve to
undefined, which would then throw an error in the spread operator. It's
therefore arguably preferable to fail immediately.

Request:None
Change-Id: I56ce81819d7a8371569b594870250f1f4edafcb4