Do not end single line comments with ;
... it makes Sonarcloud think that they are commented out code and flags
them for removal. Also, it usually does not make grammatical sense
either.

In the case of the comments in EscapeXmlCharactersResolver#escape it
would not have made sense to just remove the ';', so the whole comments
were removed instead. They are not really required to understand what is
being done anyway.

Request: None
Change-Id: I574d6c5021e58d681d52af6bf90ba55f6394d03f