Avoid loss of % sign when running bdeploy CLI
The "call" command on windows opens a new execution context
which will fiddle with the expansion of ... things.

This leads to the parameters being parsed and expanded, which
is bad in case a percent sign is inside them. Not using call
will not give control back to the outer script, but instead
pass control to the _bdeploy.bat and never return. This is
ok in this case, and parameters are handled correctly then.

Bug: CT_BDEPLOY-157
Change-Id: I41996e713f62f836a2cf00977b7f6ff852f7a3c4