Made injected values readonly
Most of them were already made readonly with the previous "readonly"
commits. This commit makes it so that all of them are readonly - not
just the private ones.
None of these have any effect on sonarcloud. This change is just for
consistency.

Also reordered the injected values to group them by their accessibility
modifier where it makes sense.

Request: None
Change-Id: Ib5485ae5e02e2536f565425e08a706a85dbff577