No longer fail on missing manifests during fetch.
It may occur that manifests with ephemeral or frequently updated
information are no longer present when the fetch operation tries
to read them during the process. They may have been when first
determining what to fetch, but no longer when trying to determine
the contents of said manifests.

In this case we now no longer fail but write a warning instead,
on either side depending on who discovers first that the manifest
is no longer there.

Bug: CT_BDEPLOY-227
Change-Id: I98fbcaecf71a26c3ff388830c50ba70721d69acc