Quick-guess searchable instance IDs for instance groups.
This *vastly* improves performance of instance group loads,
as it no longer tries to actually find and load all instance
manifests in the BHive.

Instead, this "guesses" searchable instance IDs by finding
all top-level directories in the manifest database which
match the pattern of instance IDs.

This educated guess is currently correct in 100% of the
cases. Even if we add something in the future that would
reduce this score, this would not be dramatic, as this
search "enhancement" is pure convenience.

Request: CT_BDEPLOY-128
Change-Id: I9962a22d4d1b17d71029c31d935e3b6be9bd09e5