Finally allow accessing via links with /api
We do expect /api to be present on all api-related links.
Sometimes users just click those links, or paste those URLs.
This accesses a non-existant resource, as there is none at
the root level of /api.

This commit changes this and adds a resource which simply
"redirects one up", to allow accessing the web UI when accessing
the root /api directly in the browser.

Request: CT_BDEPLOY-253
Change-Id: I01de7ba40d122d1ac66ec954e5f59110218183e9