Update text for RUNNING process in instance history.
The problem with the text is, that RUNNING can be reached
either when the process is starting up (where the old text
was ok), but also when it transitions from RUNNING_NOT_ALIVE
which can happen multiple times during a process lifetime.

The new text fits both starting up, as well as recovering
from a not-alive report.

Bug: CT_BDEPLOY-236
Change-Id: I330b6b5e345cfad8b4a5fd71140a09e36e4cdba8