Align and correct "used-in" logic for products.
First of all bulk delete had a different logic to determine
whether a product is used than single delete, which is obviously
not good.

Second, the logic missed the case where a product was used in
an instance which was not (yet) ever installed anywhere - in
this case the product should still be kept around of course.

Bug: CT_BDEPLOY-228
Change-Id: Ie90a62b0aa03415d3610f641e3cf1078f701c617