Node synchronization with more than one nodes
Issue was that all nodes simultanously tried to obtain the
same action lock (group & instance). However the *node* should
be a vital part of the action lock to reflect that the operation
is happening there. Then all nodes can work in parallel again.

The initial idea to re-use the INSTALL & ACTIVATE actions for
this turned out to be a bad one... We now introduced a new
SYNC_NODE action which is reserved solely for this. This will
now also highlight which node is synchronized, also in the UI.

Bug: CT_BDEPLOY-261
Change-Id: Ia92d14f5ab16116da74696804b644819d0f213be