Adjust remove events to be less UI centric
Until now, almost all events are sent out at least
partially by UI resources, which is wrong. However
changing this to the master resources is not straight
forward, as the CENTRAL/MANAGED interaction requires
more work to properly forward/handle events.

However, remove events are handled properly due to
the special handling in synchronizeTransacted, so
they can be moved to the correct location.

Bug: CT_BDEPLOY-246
Change-Id: I957b30f14cec8a6a8181bebf79ac946444e1e150