Cleanup: do NOT prune all hives on CENTRAL
Pruning always was a fix for rolling manifests which go undetected
in cleanup as "we" did not clean anything, but the daily business
on the server produces and deleted manifests.

However that does not happen on CENTRAL, *and* CENTRAL has a lot more
hives which creates a huge load during cleanup (e.g. one server took
more than 20 hours to run the cleanup - 2 hours before the "fix").

This change restores the previous behaviour on CENTRAL, but keeps
the "always prune" policy for non-CENTRAL

Change-Id: Iab44881ce67bad26d7cc004bf0c11ca41b2efcbe