Use arrow function for markUnresolvedExpansion
EditorComponent.markUnresolvedExpansion is passed down to another
component (BdEditorComponent) as callback.
markUnresolvedExpansion uses this.instance which does not exist on the
caller (BdEditorComponent).
Changed markUnresolvedExpansion to arrow
function so that <this> always refers to EditorComponent.
Also moved markUnresolvedExpansion higher in code, since it is not a
function anymore

Request: CT_BDEPLOY-377
Change-Id: I8f29a4a1750c900fdac5f890693555315afc651e