Change --join semantics to handle INSTANCE applications
The current behavior is to just wait for the application to exit.
This has been changed now to still wait for exit on applications
with start type MANUAL and MANUAL_CONFIRM, but not for INSTANCE,
where instead we are now waiting for starting to be completed.

Also now stopping applications is also handled so that it
can be joined.

Bug: CT_BDEPLOY-416
Change-Id: I298a321e4fd087af83d39deeab6cc243a8b25db1