Fix validation of config files in template application
In the Web UI, configuration files on new instances are not
validated. But on the backend when applying system template
or instance template (through response file), those files
were validated as if they were changed by the user.

Also improve the validation message in case any validation
issue occured while creating the instance to include app
and parameter ID in case they are set.

Bug: CT_BDEPLOY-452
Change-Id: I6d67ce284fc2ece0def13e31ed4a36550621022d