Make sure blackbird is only used for static mappers
Blackbird can cause massive creations of lambda classes
in the class metaspace, due to the way it creates and caches
anonymous optimized classes for DTOs.

This is not an issue for static mappers as those will reuse
the same generated classes over and over again, however mappers
that are frequently created will create classes with strong
references (java API limitation), and thus will never be
unlinked again.

Bug: CT_BDEPLOY-493
Change-Id: I5993f49e2888fe1e270c753bcaffb2ec981f2176