Fix issue with pressing install button too fast.
The issue is that after saving the new instance configuration
is loaded asynchronously from the server after being notified
via websocket. This can take longer than the page to load and
render - thus the install button would install the *previous*
version of the instance, not the just-saved-one.

This obviously also needs a fix, however right now we can wait
for the correct instance by matching the expected version on
screen.

Request: CT_BDEPLOY-469
Change-Id: Id22f382910f8f6b8fd7eb1f31837d5b923a9def8