Make sure to clean up *all* instance data when syncing
Up until now, on the root manifest was deleted for instances
that have been deleted on the managed server. This is a problem
as it leaves node manifests on the central. With the current
logic this should not do harm, as instance IDs are no longer
reused after deletion (this was the case in older versions).

Still it is a good idea to keep instance data consistent and
clean stuff that is no longer required

Bug: CT_BDEPLOY-729
Change-Id: I5c603914bcbbb43556e03a2d183d12f93da8f968