Fixed super tiny bug in BrowserDialogAutostartCellRenderer
The old impl would show color DISABLED if v is not instanceof Boolean,
but the correct color in that case would be COULD_NOT_CALCULATE.

This is super duper minor because v instanceof Boolean is always always
true and if it isn't we really have some kind of big issue.

Well, now it's fixed anyways.

Request: CT_BDEPLOY-691
Change-Id: I839c3b89f6913931e837f20313c7ed08231e013b