Implemented handling for node types
Completely removed the __ClientApplications constant from the web UI
EXCEPT for one unfortunate spot in system-template.components.ts where
it would be a huuuuuge effort in both back-and-frontend logic.

In the backend, quite some references to the constant had to remain in
order to keep backwards compatibility.

One noteworthy location where the constant remains is the logic
surrounding instance templates. The one instance in the frontend also
falls in this category.

Request: CT_BDEPLOY-570
Change-Id: I04640f4c39d51cd8a62eb9393dff60cbefb3d00b