Added endpoints to install/activate the newest version
Pure convenience commit. It's annoying to always have to look up the
version when testing something on the CLI.

The implementation on the backend was trivial - the logic was already
coded to do handle a null-tag in this way. Unfortunately, the tag is
part of the path of the old endpoints, so setting that to null is not
allowed. Therefore, new endpoints were needed.

Request: MO
Change-Id: Ic6fa8a14a6d701f65d6f46afccfee5264ce1b4c6